+++ desktop/js/AbeilleNetwork.js
@@ -902,7 +902,7 @@ function displayLinksGraph(gtwId) {
                                 .attr("width", imgSize)
                                 .attr("height", imgSize)
                                 .link(
-                                    "/plugins/Abeille/images/node_" +
+                                    "/plugins/Abeille/core/config/devices/devices_images/node_" +
                                         iconName +
                                         ".png"
                                 );