+++ docs/fr_FR/Changelog.md
@@ -6,6 +6,7 @@
 -   Améliorations: Assistant modèle pour device 'lightning'.
 -   Améliorations: Interne Abeille.class (publishMosquitto => msgToCmd).
 -   Nouveau: Support TRADFRIbulbE27WSglobe1055lm (2755).
+-   Améliorations: Images des devices déplacées dans 'core/config/devices/devices_images'.
 
 ## 250318-BETA-1